+++ apps/docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @js-solutions/docs
+# @zetahdev/docs
 
 ## 0.0.1
 
@@ -9,5 +9,5 @@
 - Updated dependencies [bac467b]
 - Updated dependencies [eaa471e]
 - Updated dependencies [5055747]
-  - @js-solutions/react-ui@0.1.0
-  - @js-solutions/tokens@0.1.0
+  - @zetahdev/react-ui@0.1.0
+  - @zetahdev/tokens@0.1.0